+++ src/main/java/home/example/board/domain/PostLike.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class PostLike {
-    private int post_like_seq;
+    private long post_like_seq;
     private long post_seq;
     private long user_seq;
     private String like_type; // "LIKE" or "DISLIKE"
