+++ src/main/java/home/example/board/domain/Subject.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 @Data
 public class Subject {
-    private int subject_seq;
+    private long subject_seq;
     private String subject_name;
     private LocalDateTime insert_ts;
 }
